Add four main sections to the homepage
We added 4 main sections to the homepage for the user to navigate
through the site, each section will link to a page of the site.

We use Bulma's column system to allow each section to be responsive,
also added some breakpoints for small and normal desktops. We
builded the sections mobile first.